Move link creation
